requirements.txt
setuptools==68.2.2
pylint==3.0.1
argparse==1.4.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
boto3==1.28.62